Add files via upload
You test by installing the apk.
In the app, the user can input all the fields as stated into the question and as he presses the "create child" raised button , a modal shows up to verify the details. After this, the user can view the tree structure by pressing the "view tree" button (blue coloured).